@@ -16,10 +16,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 
-version: 1.0.0-alpha+149
+version: 1.0.0-alpha+150
 
 environment:
-  sdk: '>=2.16.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -69,7 +69,7 @@ dependencies:
   path_provider: ^2.0.9
 
   # MAP
-  location: ^4.3.0
+  #location: ^4.3.0
   geolocator: ^8.2.1
 #  webview_flutter: ^3.0.0
 
@@ -107,6 +107,7 @@ flutter:
     - assets/backgrounds/
     - assets/animations/rive/
     - assets/animations/lottie/
+    - assets/img/logo/
 
 
   fonts:
