@@ -26,7 +26,6 @@ class MessageLookup extends MessageLookupByLibrary {
   static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
         "about": MessageLookupByLibrary.simpleMessage("সম্পর্কে"),
         "appLang": MessageLookupByLibrary.simpleMessage("অ্যাপ্লিকেশনের ভাষা"),
-        "asr": MessageLookupByLibrary.simpleMessage("Asr"),
         "backToHomePage":
             MessageLookupByLibrary.simpleMessage("হোমপেজে ব্যাক করো"),
         "backendError": MessageLookupByLibrary.simpleMessage(
@@ -51,12 +50,10 @@ class MessageLookup extends MessageLookupByLibrary {
         "drawerDesc": MessageLookupByLibrary.simpleMessage(
             "Connecting Muslims to Mosques"),
         "drawerTitle": MessageLookupByLibrary.simpleMessage("MAWAQIT"),
-        "duhr": MessageLookupByLibrary.simpleMessage("Duhr"),
         "enterMosqueId":
             MessageLookupByLibrary.simpleMessage("মসজিদ আইডি লিখুন:"),
         "facebook": MessageLookupByLibrary.simpleMessage(
             "ফেসবুকের মাধ্যমে সাইন আপ করুন"),
-        "fajr": MessageLookupByLibrary.simpleMessage("Fajr"),
         "gpsError": MessageLookupByLibrary.simpleMessage(
             "বর্তমান ডিভাইস অবস্থান অ্যাক্সেস করতে পারবেন না, আপনার ডিভাইস GPS সক্ষম আছে কিনা দয়া করে পরীক্ষা করুন"),
         "headerType": MessageLookupByLibrary.simpleMessage("Header type"),
@@ -64,15 +61,12 @@ class MessageLookup extends MessageLookupByLibrary {
         "instagram": MessageLookupByLibrary.simpleMessage("Instagram"),
         "installationGuide":
             MessageLookupByLibrary.simpleMessage("ইনস্টলকরণের সহায়িকা"),
-        "isha": MessageLookupByLibrary.simpleMessage("Isha"),
-        "jumua": MessageLookupByLibrary.simpleMessage("Jumua"),
         "languages": MessageLookupByLibrary.simpleMessage("ভাষা"),
         "leftButtonOption":
             MessageLookupByLibrary.simpleMessage("Radio Button Options (left)"),
         "lightMode": MessageLookupByLibrary.simpleMessage("হালকা মোড"),
         "loadingAnimation":
             MessageLookupByLibrary.simpleMessage("Loading animation"),
-        "maghrib": MessageLookupByLibrary.simpleMessage("Maghrib"),
         "mawaqitDesc": MessageLookupByLibrary.simpleMessage(
             "মাওয়াকিট আপনাকে নামাজের সময়গুলি ট্র্যাক এবং পরিচালনা করার একটি নতুন উপায় অফার করে, প্রকৃতপক্ষে আমরা একটি এন্ড-টু-এন্ড সিস্টেম অফার করি যা মসজিদ পরিচালকদের 24/24 ঘন্টা উপলব্ধ একটি অনলাইন টুল সরবরাহ করে।"),
         "mawaqitWelcome":
@@ -81,41 +75,19 @@ class MessageLookup extends MessageLookupByLibrary {
         "missingMosqueId":
             MessageLookupByLibrary.simpleMessage("মসজিদের আইডি নেই"),
         "mosque": MessageLookupByLibrary.simpleMessage("মসজিদ"),
-        "mosqueAnnouncement": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
         "mosqueId": MessageLookupByLibrary.simpleMessage("মসজিদ আইডি "),
         "mosqueIdIsNotValid": m0,
         "mosqueInput": MessageLookupByLibrary.simpleMessage("Mosque Input"),
         "mosqueNameError":
             MessageLookupByLibrary.simpleMessage("মসজিদের নাম লিখুন"),
-        "mosqueNoMore": MessageLookupByLibrary.simpleMessage("No more results"),
-        "mosqueNoResults": MessageLookupByLibrary.simpleMessage("No results"),
-        "mosque_Facilities_AblutionsRoom":
-            MessageLookupByLibrary.simpleMessage("Ablution room"),
-        "mosque_Facilities_AdultsCourse":
-            MessageLookupByLibrary.simpleMessage("Adults courses"),
-        "mosque_Facilities_ChildrenCourses":
-            MessageLookupByLibrary.simpleMessage("Children courses"),
-        "mosque_Facilities_DisabledAccess":
-            MessageLookupByLibrary.simpleMessage("Disabled access"),
-        "mosque_Facilities_Parking":
-            MessageLookupByLibrary.simpleMessage("Parking"),
-        "mosque_Facilities_SalatAlAid":
-            MessageLookupByLibrary.simpleMessage("Salat Al Eid"),
-        "mosque_Facilities_SalatAlJanaza":
-            MessageLookupByLibrary.simpleMessage("Salat Al janaza"),
-        "mosque_Facilities_WomanSpace":
-            MessageLookupByLibrary.simpleMessage("Woman space"),
-        "mosque_facilities_Info": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
-        "mosques_Facilities_IftarRamadan":
-            MessageLookupByLibrary.simpleMessage("Iftar Ramadan"),
+        "mosqueNoMore":
+            MessageLookupByLibrary.simpleMessage("আরো কোনো ফলাফল নেই"),
+        "mosqueNoResults": MessageLookupByLibrary.simpleMessage("রেসাল্টস নেই"),
         "navigationBarStyle":
             MessageLookupByLibrary.simpleMessage("Navigation bars style"),
         "no": MessageLookupByLibrary.simpleMessage("না"),
         "noInternet":
             MessageLookupByLibrary.simpleMessage("কোনও ইন্টারনেট সংযোগ নেই"),
-        "noJumua": MessageLookupByLibrary.simpleMessage("no jumua"),
         "notification": MessageLookupByLibrary.simpleMessage("নোটিফিকেশন সমূহ"),
         "ok": MessageLookupByLibrary.simpleMessage("আচ্ছা"),
         "privacyPolicy":
@@ -132,7 +104,6 @@ class MessageLookup extends MessageLookupByLibrary {
         "selectWithMosqueId": MessageLookupByLibrary.simpleMessage(
             "অনুগ্রহ করে আপনার \'মসজিদ আইডি\' লিখুন"),
         "share": MessageLookupByLibrary.simpleMessage("শেয়ার"),
-        "shuruq": MessageLookupByLibrary.simpleMessage("shuruq"),
         "skype": MessageLookupByLibrary.simpleMessage("স্কাইপ"),
         "slugError": MessageLookupByLibrary.simpleMessage(
             "একটি বৈধ মসজিদ \'স্লাগ\' নয়"),
