@@ -26,7 +26,6 @@ class MessageLookup extends MessageLookupByLibrary {
   static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
         "about": MessageLookupByLibrary.simpleMessage("Über"),
         "appLang": MessageLookupByLibrary.simpleMessage("App-Sprache"),
-        "asr": MessageLookupByLibrary.simpleMessage("Asr"),
         "backToHomePage":
             MessageLookupByLibrary.simpleMessage("Zurück zur Startseite"),
         "backendError": MessageLookupByLibrary.simpleMessage(
@@ -47,11 +46,9 @@ class MessageLookup extends MessageLookupByLibrary {
         "drawerDesc": MessageLookupByLibrary.simpleMessage(
             "Connecting Muslims to Mosques"),
         "drawerTitle": MessageLookupByLibrary.simpleMessage("MAWAQIT"),
-        "duhr": MessageLookupByLibrary.simpleMessage("Duhr"),
         "enterMosqueId":
             MessageLookupByLibrary.simpleMessage("Moschee ID eingeben: "),
         "facebook": MessageLookupByLibrary.simpleMessage("Facebook"),
-        "fajr": MessageLookupByLibrary.simpleMessage("Fajr"),
         "gpsError": MessageLookupByLibrary.simpleMessage(
             "Zugriff auf den aktuellen Gerätespeicher nicht möglich, bitte überprüfen Sie, ob Ihr Gerät GPS aktiviert ist"),
         "headerType": MessageLookupByLibrary.simpleMessage("Header-Typ"),
@@ -59,15 +56,12 @@ class MessageLookup extends MessageLookupByLibrary {
         "instagram": MessageLookupByLibrary.simpleMessage("Instagram"),
         "installationGuide":
             MessageLookupByLibrary.simpleMessage("Installationsanleitungen"),
-        "isha": MessageLookupByLibrary.simpleMessage("Isha"),
-        "jumua": MessageLookupByLibrary.simpleMessage("Jumua"),
         "languages": MessageLookupByLibrary.simpleMessage("Sprachen"),
         "leftButtonOption": MessageLookupByLibrary.simpleMessage(
             "Optionen der Optionsleiste (links)"),
         "lightMode": MessageLookupByLibrary.simpleMessage("Lichtmodus"),
         "loadingAnimation":
             MessageLookupByLibrary.simpleMessage("Lade Animation"),
-        "maghrib": MessageLookupByLibrary.simpleMessage("Maghrib"),
         "mawaqitDesc": MessageLookupByLibrary.simpleMessage(
             "MAWAQIT bietet dir eine neue Möglichkeit, Gebetszeiten zu verfolgen und zu verwalten, in der Tat bieten wir ein Ende-zu-End-System, das Mosche-Managern ein Online-Tool rund um die Uhr zur Verfügung stellt."),
         "mawaqitWelcome":
@@ -76,41 +70,20 @@ class MessageLookup extends MessageLookupByLibrary {
         "missingMosqueId":
             MessageLookupByLibrary.simpleMessage("Fehlende Mosche ID"),
         "mosque": MessageLookupByLibrary.simpleMessage("Moschee"),
-        "mosqueAnnouncement": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
         "mosqueId": MessageLookupByLibrary.simpleMessage("Moschee ID "),
         "mosqueIdIsNotValid": m0,
         "mosqueInput": MessageLookupByLibrary.simpleMessage("Moschee-Eingabe"),
         "mosqueNameError":
             MessageLookupByLibrary.simpleMessage("Namen der Moschee eingeben"),
-        "mosqueNoMore": MessageLookupByLibrary.simpleMessage("No more results"),
-        "mosqueNoResults": MessageLookupByLibrary.simpleMessage("No results"),
-        "mosque_Facilities_AblutionsRoom":
-            MessageLookupByLibrary.simpleMessage("Ablution room"),
-        "mosque_Facilities_AdultsCourse":
-            MessageLookupByLibrary.simpleMessage("Adults courses"),
-        "mosque_Facilities_ChildrenCourses":
-            MessageLookupByLibrary.simpleMessage("Children courses"),
-        "mosque_Facilities_DisabledAccess":
-            MessageLookupByLibrary.simpleMessage("Disabled access"),
-        "mosque_Facilities_Parking":
-            MessageLookupByLibrary.simpleMessage("Parking"),
-        "mosque_Facilities_SalatAlAid":
-            MessageLookupByLibrary.simpleMessage("Salat Al Eid"),
-        "mosque_Facilities_SalatAlJanaza":
-            MessageLookupByLibrary.simpleMessage("Salat Al janaza"),
-        "mosque_Facilities_WomanSpace":
-            MessageLookupByLibrary.simpleMessage("Woman space"),
-        "mosque_facilities_Info": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
-        "mosques_Facilities_IftarRamadan":
-            MessageLookupByLibrary.simpleMessage("Iftar Ramadan"),
+        "mosqueNoMore":
+            MessageLookupByLibrary.simpleMessage("Keine weiteren Ergebnisse"),
+        "mosqueNoResults":
+            MessageLookupByLibrary.simpleMessage("Kein Ergebnis"),
         "navigationBarStyle":
             MessageLookupByLibrary.simpleMessage("Navigationsleisten Stil"),
         "no": MessageLookupByLibrary.simpleMessage("Nein"),
         "noInternet":
             MessageLookupByLibrary.simpleMessage("Keine Internetverbindung"),
-        "noJumua": MessageLookupByLibrary.simpleMessage("no jumua"),
         "notification":
             MessageLookupByLibrary.simpleMessage("Benachrichtigungen"),
         "ok": MessageLookupByLibrary.simpleMessage("Ok"),
@@ -128,7 +101,6 @@ class MessageLookup extends MessageLookupByLibrary {
         "selectWithMosqueId": MessageLookupByLibrary.simpleMessage(
             "Bitte geben Sie Ihre Mosche-ID ein"),
         "share": MessageLookupByLibrary.simpleMessage("Teilen"),
-        "shuruq": MessageLookupByLibrary.simpleMessage("shuruq"),
         "skype": MessageLookupByLibrary.simpleMessage("Skype"),
         "slugError": MessageLookupByLibrary.simpleMessage(
             "Ist kein gültiger Mosche-Slug"),
