@@ -27,7 +27,6 @@ class MessageLookup extends MessageLookupByLibrary {
         "about": MessageLookupByLibrary.simpleMessage("Acerca de"),
         "appLang":
             MessageLookupByLibrary.simpleMessage("Idioma de la aplicación"),
-        "asr": MessageLookupByLibrary.simpleMessage("Asr"),
         "backToHomePage": MessageLookupByLibrary.simpleMessage(
             "Volver a la página de inicio"),
         "backendError": MessageLookupByLibrary.simpleMessage(
@@ -50,11 +49,9 @@ class MessageLookup extends MessageLookupByLibrary {
         "drawerDesc": MessageLookupByLibrary.simpleMessage(
             "Connecting Muslims to Mosques"),
         "drawerTitle": MessageLookupByLibrary.simpleMessage("MAWAQIT"),
-        "duhr": MessageLookupByLibrary.simpleMessage("Duhr"),
         "enterMosqueId": MessageLookupByLibrary.simpleMessage(
             "Introducir ID de Mezquita : "),
         "facebook": MessageLookupByLibrary.simpleMessage("Facebook"),
-        "fajr": MessageLookupByLibrary.simpleMessage("Fajr"),
         "gpsError": MessageLookupByLibrary.simpleMessage(
             "No se puede acceder a la ubicación actual del dispositivo, por favor comprueba que el GPS de tu dispositivo está activado"),
         "headerType": MessageLookupByLibrary.simpleMessage("Tipo de cabecera"),
@@ -62,15 +59,12 @@ class MessageLookup extends MessageLookupByLibrary {
         "instagram": MessageLookupByLibrary.simpleMessage("Instagrama"),
         "installationGuide":
             MessageLookupByLibrary.simpleMessage("Guía de Instalación"),
-        "isha": MessageLookupByLibrary.simpleMessage("Isha"),
-        "jumua": MessageLookupByLibrary.simpleMessage("Jumua"),
         "languages": MessageLookupByLibrary.simpleMessage("Idiomas"),
         "leftButtonOption": MessageLookupByLibrary.simpleMessage(
             "Opciones del botón de radio (izquierda)"),
         "lightMode": MessageLookupByLibrary.simpleMessage("Modo claro"),
         "loadingAnimation":
             MessageLookupByLibrary.simpleMessage("Cargando Animación"),
-        "maghrib": MessageLookupByLibrary.simpleMessage("Maghrib"),
         "mawaqitDesc": MessageLookupByLibrary.simpleMessage(
             "MAWAQIT te ofrece una nueva forma de rastrear y gestionar los tiempos de oración, de hecho ofrecemos un sistema de extremo a extremo que proporciona a los gestores de mezquitas una herramienta en línea disponible las 24 horas del día."),
         "mawaqitWelcome":
@@ -79,42 +73,21 @@ class MessageLookup extends MessageLookupByLibrary {
         "missingMosqueId":
             MessageLookupByLibrary.simpleMessage("Falta el ID de la mezquita"),
         "mosque": MessageLookupByLibrary.simpleMessage("Mezquita"),
-        "mosqueAnnouncement": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
         "mosqueId": MessageLookupByLibrary.simpleMessage("Mezquita ID"),
         "mosqueIdIsNotValid": m0,
         "mosqueInput":
             MessageLookupByLibrary.simpleMessage("Entrada de Mezquita"),
         "mosqueNameError": MessageLookupByLibrary.simpleMessage(
             "Introduzca el nombre de la Mezquita"),
-        "mosqueNoMore": MessageLookupByLibrary.simpleMessage("No more results"),
-        "mosqueNoResults": MessageLookupByLibrary.simpleMessage("No results"),
-        "mosque_Facilities_AblutionsRoom":
-            MessageLookupByLibrary.simpleMessage("Ablution room"),
-        "mosque_Facilities_AdultsCourse":
-            MessageLookupByLibrary.simpleMessage("Adults courses"),
-        "mosque_Facilities_ChildrenCourses":
-            MessageLookupByLibrary.simpleMessage("Children courses"),
-        "mosque_Facilities_DisabledAccess":
-            MessageLookupByLibrary.simpleMessage("Disabled access"),
-        "mosque_Facilities_Parking":
-            MessageLookupByLibrary.simpleMessage("Parking"),
-        "mosque_Facilities_SalatAlAid":
-            MessageLookupByLibrary.simpleMessage("Salat Al Eid"),
-        "mosque_Facilities_SalatAlJanaza":
-            MessageLookupByLibrary.simpleMessage("Salat Al janaza"),
-        "mosque_Facilities_WomanSpace":
-            MessageLookupByLibrary.simpleMessage("Woman space"),
-        "mosque_facilities_Info": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
-        "mosques_Facilities_IftarRamadan":
-            MessageLookupByLibrary.simpleMessage("Iftar Ramadan"),
+        "mosqueNoMore":
+            MessageLookupByLibrary.simpleMessage("No hay más resultados"),
+        "mosqueNoResults":
+            MessageLookupByLibrary.simpleMessage("No hay resultados"),
         "navigationBarStyle":
             MessageLookupByLibrary.simpleMessage("Estilo barras de navegación"),
         "no": MessageLookupByLibrary.simpleMessage("No"),
         "noInternet":
             MessageLookupByLibrary.simpleMessage("No hay conexión a internet"),
-        "noJumua": MessageLookupByLibrary.simpleMessage("no jumua"),
         "notification": MessageLookupByLibrary.simpleMessage("Notificaciones"),
         "ok": MessageLookupByLibrary.simpleMessage("Ok"),
         "privacyPolicy":
@@ -131,7 +104,6 @@ class MessageLookup extends MessageLookupByLibrary {
         "selectWithMosqueId": MessageLookupByLibrary.simpleMessage(
             "Por favor, introduzca su ID de la Mezquita"),
         "share": MessageLookupByLibrary.simpleMessage("Compartir"),
-        "shuruq": MessageLookupByLibrary.simpleMessage("shuruq"),
         "skype": MessageLookupByLibrary.simpleMessage("Skype"),
         "slugError": MessageLookupByLibrary.simpleMessage(
             "No es un slug de mezquita válido"),
