@@ -28,7 +28,6 @@ class MessageLookup extends MessageLookupByLibrary {
         "about": MessageLookupByLibrary.simpleMessage("À propos"),
         "appLang":
             MessageLookupByLibrary.simpleMessage("Langue de l\'application"),
-        "asr": MessageLookupByLibrary.simpleMessage("Asr"),
         "backToHomePage": MessageLookupByLibrary.simpleMessage(
             "Revenir à la page d\'accueil"),
         "backendError": MessageLookupByLibrary.simpleMessage(
@@ -52,11 +51,9 @@ class MessageLookup extends MessageLookupByLibrary {
         "drawerDesc": MessageLookupByLibrary.simpleMessage(
             "Connecting muslims to Mosques"),
         "drawerTitle": MessageLookupByLibrary.simpleMessage("MAWAQIT"),
-        "duhr": MessageLookupByLibrary.simpleMessage("Duhr"),
         "enterMosqueId": MessageLookupByLibrary.simpleMessage(
             "Veuillez saisir l\'ID de votre installation ou l\'ID de la mosquée (Mosque ID)"),
         "facebook": MessageLookupByLibrary.simpleMessage("Facebook"),
-        "fajr": MessageLookupByLibrary.simpleMessage("Fajr"),
         "gpsError": MessageLookupByLibrary.simpleMessage(
             "Impossible d\'accéder à votre localisation, vérifiez que le GPS de votre appareil est actif."),
         "headerType": MessageLookupByLibrary.simpleMessage("Type d\'en-tête"),
@@ -64,15 +61,12 @@ class MessageLookup extends MessageLookupByLibrary {
         "instagram": MessageLookupByLibrary.simpleMessage("Instagram"),
         "installationGuide":
             MessageLookupByLibrary.simpleMessage("Guide d\'installation"),
-        "isha": MessageLookupByLibrary.simpleMessage("Isha"),
-        "jumua": MessageLookupByLibrary.simpleMessage("Jumua"),
         "languages": MessageLookupByLibrary.simpleMessage("Langues"),
         "leftButtonOption": MessageLookupByLibrary.simpleMessage(
             "Options du bouton radio (à gauche)"),
         "lightMode": MessageLookupByLibrary.simpleMessage("Mode clair"),
         "loadingAnimation":
             MessageLookupByLibrary.simpleMessage("Animation lors du lancement"),
-        "maghrib": MessageLookupByLibrary.simpleMessage("Maghrib"),
         "mawaqitDesc": MessageLookupByLibrary.simpleMessage(
             "MAWAQIT vous offre une nouvelle façon de suivre et de gérer les temps de prière, en effet nous offrons un système de bout en bout qui fournit aux gestionnaires de mosquées un outil en ligne disponible 24/24h."),
         "mawaqitWelcome":
@@ -81,41 +75,20 @@ class MessageLookup extends MessageLookupByLibrary {
         "missingMosqueId": MessageLookupByLibrary.simpleMessage(
             "Il manque l\'ID de la mosquée"),
         "mosque": MessageLookupByLibrary.simpleMessage("Mosquée"),
-        "mosqueAnnouncement": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
         "mosqueId": MessageLookupByLibrary.simpleMessage("ID de la mosquée "),
         "mosqueIdIsNotValid": m0,
         "mosqueInput": MessageLookupByLibrary.simpleMessage("Mosque Input"),
         "mosqueNameError":
             MessageLookupByLibrary.simpleMessage("Entrer le nom de la mosquée"),
-        "mosqueNoMore": MessageLookupByLibrary.simpleMessage("No more results"),
-        "mosqueNoResults": MessageLookupByLibrary.simpleMessage("No results"),
-        "mosque_Facilities_AblutionsRoom":
-            MessageLookupByLibrary.simpleMessage("Ablution room"),
-        "mosque_Facilities_AdultsCourse":
-            MessageLookupByLibrary.simpleMessage("Adults courses"),
-        "mosque_Facilities_ChildrenCourses":
-            MessageLookupByLibrary.simpleMessage("Children courses"),
-        "mosque_Facilities_DisabledAccess":
-            MessageLookupByLibrary.simpleMessage("Disabled access"),
-        "mosque_Facilities_Parking":
-            MessageLookupByLibrary.simpleMessage("Parking"),
-        "mosque_Facilities_SalatAlAid":
-            MessageLookupByLibrary.simpleMessage("Salat Al Eid"),
-        "mosque_Facilities_SalatAlJanaza":
-            MessageLookupByLibrary.simpleMessage("Salat Al janaza"),
-        "mosque_Facilities_WomanSpace":
-            MessageLookupByLibrary.simpleMessage("Woman space"),
-        "mosque_facilities_Info": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
-        "mosques_Facilities_IftarRamadan":
-            MessageLookupByLibrary.simpleMessage("Iftar Ramadan"),
+        "mosqueNoMore":
+            MessageLookupByLibrary.simpleMessage("Aucun autre résultat"),
+        "mosqueNoResults":
+            MessageLookupByLibrary.simpleMessage("Aucun résultat"),
         "navigationBarStyle": MessageLookupByLibrary.simpleMessage(
             "Style des barres de navigation"),
         "no": MessageLookupByLibrary.simpleMessage("Non"),
         "noInternet":
             MessageLookupByLibrary.simpleMessage("Aucune connexion internet"),
-        "noJumua": MessageLookupByLibrary.simpleMessage("no jumua"),
         "notification": MessageLookupByLibrary.simpleMessage("Notifications"),
         "ok": MessageLookupByLibrary.simpleMessage("OK"),
         "privacyPolicy": MessageLookupByLibrary.simpleMessage(
@@ -132,7 +105,6 @@ class MessageLookup extends MessageLookupByLibrary {
         "selectWithMosqueId": MessageLookupByLibrary.simpleMessage(
             "Veuillez saisir l\'ID de votre installation de la mosquée (Mosque ID)"),
         "share": MessageLookupByLibrary.simpleMessage("Partager"),
-        "shuruq": MessageLookupByLibrary.simpleMessage("shuruq"),
         "skype": MessageLookupByLibrary.simpleMessage("Skype"),
         "slugError": MessageLookupByLibrary.simpleMessage(
             "Ce n\'est pas un slug de mosquée valide"),
