@@ -26,7 +26,6 @@ class MessageLookup extends MessageLookupByLibrary {
   static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
         "about": MessageLookupByLibrary.simpleMessage("Informatie"),
         "appLang": MessageLookupByLibrary.simpleMessage("App taal"),
-        "asr": MessageLookupByLibrary.simpleMessage("Asr"),
         "backToHomePage":
             MessageLookupByLibrary.simpleMessage("Terug naar Startpagina"),
         "backendError": MessageLookupByLibrary.simpleMessage(
@@ -47,11 +46,9 @@ class MessageLookup extends MessageLookupByLibrary {
         "drawerDesc": MessageLookupByLibrary.simpleMessage(
             "Connecting Muslims to Mosques"),
         "drawerTitle": MessageLookupByLibrary.simpleMessage("MAWAQIT"),
-        "duhr": MessageLookupByLibrary.simpleMessage("Duhr"),
         "enterMosqueId":
             MessageLookupByLibrary.simpleMessage("Voer Moske-ID in: "),
         "facebook": MessageLookupByLibrary.simpleMessage("Facebook"),
-        "fajr": MessageLookupByLibrary.simpleMessage("Fajr"),
         "gpsError": MessageLookupByLibrary.simpleMessage(
             "Geen toegang tot de huidige locatie van het apparaat, controleer of de GPS is ingeschakeld"),
         "headerType": MessageLookupByLibrary.simpleMessage("Header type"),
@@ -59,15 +56,12 @@ class MessageLookup extends MessageLookupByLibrary {
         "instagram": MessageLookupByLibrary.simpleMessage("Instagram"),
         "installationGuide":
             MessageLookupByLibrary.simpleMessage("Installatiehandleiding"),
-        "isha": MessageLookupByLibrary.simpleMessage("Isha"),
-        "jumua": MessageLookupByLibrary.simpleMessage("Jumua"),
         "languages": MessageLookupByLibrary.simpleMessage("Leren"),
         "leftButtonOption":
             MessageLookupByLibrary.simpleMessage("Radio Knop opties (links)"),
         "lightMode": MessageLookupByLibrary.simpleMessage("Lichte modus"),
         "loadingAnimation":
             MessageLookupByLibrary.simpleMessage("Animatie laden"),
-        "maghrib": MessageLookupByLibrary.simpleMessage("Maghrib"),
         "mawaqitDesc": MessageLookupByLibrary.simpleMessage(
             "MAWAQIT biedt je een nieuwe manier om gebedstijden bij te houden en te beheren. We bieden inderdaad een end-to-end systeem dat moskeemanagers een online tool biedt die 24/24 uur beschikbaar is."),
         "mawaqitWelcome":
@@ -76,41 +70,20 @@ class MessageLookup extends MessageLookupByLibrary {
         "missingMosqueId":
             MessageLookupByLibrary.simpleMessage("Ontbrekende moske-ID"),
         "mosque": MessageLookupByLibrary.simpleMessage("Moskee"),
-        "mosqueAnnouncement": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
         "mosqueId": MessageLookupByLibrary.simpleMessage("Moske-ID "),
         "mosqueIdIsNotValid": m0,
         "mosqueInput": MessageLookupByLibrary.simpleMessage("Moskee input"),
         "mosqueNameError":
             MessageLookupByLibrary.simpleMessage("Ga de Moske-naam in"),
-        "mosqueNoMore": MessageLookupByLibrary.simpleMessage("No more results"),
-        "mosqueNoResults": MessageLookupByLibrary.simpleMessage("No results"),
-        "mosque_Facilities_AblutionsRoom":
-            MessageLookupByLibrary.simpleMessage("Ablution room"),
-        "mosque_Facilities_AdultsCourse":
-            MessageLookupByLibrary.simpleMessage("Adults courses"),
-        "mosque_Facilities_ChildrenCourses":
-            MessageLookupByLibrary.simpleMessage("Children courses"),
-        "mosque_Facilities_DisabledAccess":
-            MessageLookupByLibrary.simpleMessage("Disabled access"),
-        "mosque_Facilities_Parking":
-            MessageLookupByLibrary.simpleMessage("Parking"),
-        "mosque_Facilities_SalatAlAid":
-            MessageLookupByLibrary.simpleMessage("Salat Al Eid"),
-        "mosque_Facilities_SalatAlJanaza":
-            MessageLookupByLibrary.simpleMessage("Salat Al janaza"),
-        "mosque_Facilities_WomanSpace":
-            MessageLookupByLibrary.simpleMessage("Woman space"),
-        "mosque_facilities_Info": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
-        "mosques_Facilities_IftarRamadan":
-            MessageLookupByLibrary.simpleMessage("Iftar Ramadan"),
+        "mosqueNoMore":
+            MessageLookupByLibrary.simpleMessage("Niet meer resultaten"),
+        "mosqueNoResults":
+            MessageLookupByLibrary.simpleMessage("Geen resultaten"),
         "navigationBarStyle":
             MessageLookupByLibrary.simpleMessage("Stijl navigatiebalken"),
         "no": MessageLookupByLibrary.simpleMessage("Nee"),
         "noInternet":
             MessageLookupByLibrary.simpleMessage("Geen internetverbinding"),
-        "noJumua": MessageLookupByLibrary.simpleMessage("no jumua"),
         "notification": MessageLookupByLibrary.simpleMessage("Meldingen"),
         "ok": MessageLookupByLibrary.simpleMessage("OK"),
         "privacyPolicy": MessageLookupByLibrary.simpleMessage("Privacybeleid"),
@@ -125,7 +98,6 @@ class MessageLookup extends MessageLookupByLibrary {
         "selectWithMosqueId":
             MessageLookupByLibrary.simpleMessage("Voer Moske-ID in"),
         "share": MessageLookupByLibrary.simpleMessage("Delen"),
-        "shuruq": MessageLookupByLibrary.simpleMessage("shuruq"),
         "skype": MessageLookupByLibrary.simpleMessage("Skype"),
         "slugError": MessageLookupByLibrary.simpleMessage(
             "Dit is geen geldige moskee slug"),
