@@ -26,7 +26,6 @@ class MessageLookup extends MessageLookupByLibrary {
   static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
         "about": MessageLookupByLibrary.simpleMessage("SOBRE"),
         "appLang": MessageLookupByLibrary.simpleMessage("Idioma do aplicativo"),
-        "asr": MessageLookupByLibrary.simpleMessage("Asr"),
         "backToHomePage": MessageLookupByLibrary.simpleMessage(
             "Voltar para a página inicial"),
         "backendError": MessageLookupByLibrary.simpleMessage(
@@ -49,11 +48,9 @@ class MessageLookup extends MessageLookupByLibrary {
         "drawerDesc": MessageLookupByLibrary.simpleMessage(
             "Connecting Muslims to Mosques"),
         "drawerTitle": MessageLookupByLibrary.simpleMessage("MAWAQIT"),
-        "duhr": MessageLookupByLibrary.simpleMessage("Duhr"),
         "enterMosqueId":
             MessageLookupByLibrary.simpleMessage("Insira a ID da Mesquita: "),
         "facebook": MessageLookupByLibrary.simpleMessage("Siga-nos"),
-        "fajr": MessageLookupByLibrary.simpleMessage("Fajr"),
         "gpsError": MessageLookupByLibrary.simpleMessage(
             "Não é possível acessar a localização atual do dispositivo, por favor, verifique se o GPS do seu dispositivo está ativado"),
         "headerType": MessageLookupByLibrary.simpleMessage("Tipo de cabeçalho"),
@@ -61,15 +58,12 @@ class MessageLookup extends MessageLookupByLibrary {
         "instagram": MessageLookupByLibrary.simpleMessage("Instagram"),
         "installationGuide":
             MessageLookupByLibrary.simpleMessage("Guia de instalação"),
-        "isha": MessageLookupByLibrary.simpleMessage("Isha"),
-        "jumua": MessageLookupByLibrary.simpleMessage("Jumua"),
         "languages": MessageLookupByLibrary.simpleMessage("Idiomas"),
         "leftButtonOption": MessageLookupByLibrary.simpleMessage(
             "Opções do Botão Rádio (esquerda)"),
         "lightMode": MessageLookupByLibrary.simpleMessage("Modo claro"),
         "loadingAnimation":
             MessageLookupByLibrary.simpleMessage("Carregando Animação"),
-        "maghrib": MessageLookupByLibrary.simpleMessage("Maghrib"),
         "mawaqitDesc": MessageLookupByLibrary.simpleMessage(
             "MAWAQIT oferece uma nova maneira de acompanhar e gerenciar as preces, na verdade, oferecemos um sistema de ponta a ponta que oferece aos gestores de mesquitas uma ferramenta online disponível 24/24h."),
         "mawaqitWelcome":
@@ -78,42 +72,21 @@ class MessageLookup extends MessageLookupByLibrary {
         "missingMosqueId":
             MessageLookupByLibrary.simpleMessage("ID da mesquita ausente"),
         "mosque": MessageLookupByLibrary.simpleMessage("Mesquita"),
-        "mosqueAnnouncement": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
         "mosqueId": MessageLookupByLibrary.simpleMessage("ID de Mesquita "),
         "mosqueIdIsNotValid": m0,
         "mosqueInput":
             MessageLookupByLibrary.simpleMessage("Entrada de Mesquita"),
         "mosqueNameError":
             MessageLookupByLibrary.simpleMessage("Digite o nome da Mesquita"),
-        "mosqueNoMore": MessageLookupByLibrary.simpleMessage("No more results"),
-        "mosqueNoResults": MessageLookupByLibrary.simpleMessage("No results"),
-        "mosque_Facilities_AblutionsRoom":
-            MessageLookupByLibrary.simpleMessage("Ablution room"),
-        "mosque_Facilities_AdultsCourse":
-            MessageLookupByLibrary.simpleMessage("Adults courses"),
-        "mosque_Facilities_ChildrenCourses":
-            MessageLookupByLibrary.simpleMessage("Children courses"),
-        "mosque_Facilities_DisabledAccess":
-            MessageLookupByLibrary.simpleMessage("Disabled access"),
-        "mosque_Facilities_Parking":
-            MessageLookupByLibrary.simpleMessage("Parking"),
-        "mosque_Facilities_SalatAlAid":
-            MessageLookupByLibrary.simpleMessage("Salat Al Eid"),
-        "mosque_Facilities_SalatAlJanaza":
-            MessageLookupByLibrary.simpleMessage("Salat Al janaza"),
-        "mosque_Facilities_WomanSpace":
-            MessageLookupByLibrary.simpleMessage("Woman space"),
-        "mosque_facilities_Info": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
-        "mosques_Facilities_IftarRamadan":
-            MessageLookupByLibrary.simpleMessage("Iftar Ramadan"),
+        "mosqueNoMore":
+            MessageLookupByLibrary.simpleMessage("Sem mais resultados"),
+        "mosqueNoResults":
+            MessageLookupByLibrary.simpleMessage("Sem resultados"),
         "navigationBarStyle": MessageLookupByLibrary.simpleMessage(
             "Estilo das barras de navegação"),
         "no": MessageLookupByLibrary.simpleMessage("Não"),
         "noInternet":
             MessageLookupByLibrary.simpleMessage("Sem conexão à internet"),
-        "noJumua": MessageLookupByLibrary.simpleMessage("no jumua"),
         "notification": MessageLookupByLibrary.simpleMessage("Notificações"),
         "ok": MessageLookupByLibrary.simpleMessage("CERTO"),
         "privacyPolicy":
@@ -130,7 +103,6 @@ class MessageLookup extends MessageLookupByLibrary {
         "selectWithMosqueId": MessageLookupByLibrary.simpleMessage(
             "Por favor, insira sua ID de Mesquita"),
         "share": MessageLookupByLibrary.simpleMessage("Compartilhar"),
-        "shuruq": MessageLookupByLibrary.simpleMessage("shuruq"),
         "skype": MessageLookupByLibrary.simpleMessage("Skype"),
         "slugError": MessageLookupByLibrary.simpleMessage(
             "Não é um slug de mesquita válido"),
