@@ -26,7 +26,6 @@ class MessageLookup extends MessageLookupByLibrary {
   static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
         "about": MessageLookupByLibrary.simpleMessage("Hakkında"),
         "appLang": MessageLookupByLibrary.simpleMessage("Uygulama Dili"),
-        "asr": MessageLookupByLibrary.simpleMessage("Asr"),
         "backToHomePage":
             MessageLookupByLibrary.simpleMessage("Ana sayfaya dön"),
         "backendError": MessageLookupByLibrary.simpleMessage(
@@ -47,11 +46,9 @@ class MessageLookup extends MessageLookupByLibrary {
         "drawerDesc": MessageLookupByLibrary.simpleMessage(
             "Connecting Muslims to Mosques"),
         "drawerTitle": MessageLookupByLibrary.simpleMessage("MAWAQIT"),
-        "duhr": MessageLookupByLibrary.simpleMessage("Duhr"),
         "enterMosqueId": MessageLookupByLibrary.simpleMessage(
             "Lütfen Cami Kimliğinizi giriniz : "),
         "facebook": MessageLookupByLibrary.simpleMessage("Facebook"),
-        "fajr": MessageLookupByLibrary.simpleMessage("Fajr"),
         "gpsError": MessageLookupByLibrary.simpleMessage(
             "Mevcut cihaz konumuna erişilemiyor, lütfen cihazınızın GPS\'inin etkin olduğunu kontrol edin"),
         "headerType": MessageLookupByLibrary.simpleMessage("Başlık Türü"),
@@ -59,15 +56,12 @@ class MessageLookup extends MessageLookupByLibrary {
         "instagram": MessageLookupByLibrary.simpleMessage("Instagram"),
         "installationGuide":
             MessageLookupByLibrary.simpleMessage("Kurulum Kılavuzu"),
-        "isha": MessageLookupByLibrary.simpleMessage("Isha"),
-        "jumua": MessageLookupByLibrary.simpleMessage("Jumua"),
         "languages": MessageLookupByLibrary.simpleMessage("Diller"),
         "leftButtonOption":
             MessageLookupByLibrary.simpleMessage("Radio Button Options (left)"),
         "lightMode": MessageLookupByLibrary.simpleMessage("Aydınlık mod"),
         "loadingAnimation":
             MessageLookupByLibrary.simpleMessage("Şarj animasyonu"),
-        "maghrib": MessageLookupByLibrary.simpleMessage("Maghrib"),
         "mawaqitDesc": MessageLookupByLibrary.simpleMessage(
             "MAWAQIT size namaz vakitlerini takip etmeniz ve yönetmeniz için yeni bir yol sunuyor, aslında biz cami yöneticilerine 24/24 saat çevrimiçi bir araç sağlayan uçtan uca bir sistem sunuyoruz."),
         "mawaqitWelcome":
@@ -76,41 +70,19 @@ class MessageLookup extends MessageLookupByLibrary {
         "missingMosqueId":
             MessageLookupByLibrary.simpleMessage("Eksik Cami ID"),
         "mosque": MessageLookupByLibrary.simpleMessage("Cami"),
-        "mosqueAnnouncement": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
         "mosqueId": MessageLookupByLibrary.simpleMessage("Cami ID (sayı) "),
         "mosqueIdIsNotValid": m0,
         "mosqueInput": MessageLookupByLibrary.simpleMessage("Cami Girişi"),
         "mosqueNameError":
             MessageLookupByLibrary.simpleMessage("Cami adını girin"),
-        "mosqueNoMore": MessageLookupByLibrary.simpleMessage("No more results"),
-        "mosqueNoResults": MessageLookupByLibrary.simpleMessage("No results"),
-        "mosque_Facilities_AblutionsRoom":
-            MessageLookupByLibrary.simpleMessage("Ablution room"),
-        "mosque_Facilities_AdultsCourse":
-            MessageLookupByLibrary.simpleMessage("Adults courses"),
-        "mosque_Facilities_ChildrenCourses":
-            MessageLookupByLibrary.simpleMessage("Children courses"),
-        "mosque_Facilities_DisabledAccess":
-            MessageLookupByLibrary.simpleMessage("Disabled access"),
-        "mosque_Facilities_Parking":
-            MessageLookupByLibrary.simpleMessage("Parking"),
-        "mosque_Facilities_SalatAlAid":
-            MessageLookupByLibrary.simpleMessage("Salat Al Eid"),
-        "mosque_Facilities_SalatAlJanaza":
-            MessageLookupByLibrary.simpleMessage("Salat Al janaza"),
-        "mosque_Facilities_WomanSpace":
-            MessageLookupByLibrary.simpleMessage("Woman space"),
-        "mosque_facilities_Info": MessageLookupByLibrary.simpleMessage(
-            "All this information is filled in by the people in charge of the mosque"),
-        "mosques_Facilities_IftarRamadan":
-            MessageLookupByLibrary.simpleMessage("Iftar Ramadan"),
+        "mosqueNoMore":
+            MessageLookupByLibrary.simpleMessage("Daha fazla sonuç yok"),
+        "mosqueNoResults": MessageLookupByLibrary.simpleMessage("Sonuç yok"),
         "navigationBarStyle": MessageLookupByLibrary.simpleMessage(
             "Gezinme çubuğu görüntüleme stili"),
         "no": MessageLookupByLibrary.simpleMessage("Hayır"),
         "noInternet":
             MessageLookupByLibrary.simpleMessage("İnternet bağlantısı yok"),
-        "noJumua": MessageLookupByLibrary.simpleMessage("no jumua"),
         "notification": MessageLookupByLibrary.simpleMessage("Bildirimler"),
         "ok": MessageLookupByLibrary.simpleMessage("TAMAM"),
         "privacyPolicy":
@@ -125,7 +97,6 @@ class MessageLookup extends MessageLookupByLibrary {
         "selectWithMosqueId": MessageLookupByLibrary.simpleMessage(
             "Lütfen Cami Kimliğinizi giriniz (Mosque ID)"),
         "share": MessageLookupByLibrary.simpleMessage("Paylaş"),
-        "shuruq": MessageLookupByLibrary.simpleMessage("shuruq"),
         "skype": MessageLookupByLibrary.simpleMessage("Skype"),
         "slugError": MessageLookupByLibrary.simpleMessage(
             "Geçerli bir cami \'slug\' değil mi"),
