@@ -1,10 +1,10 @@
 import 'dart:convert';
 
 import 'package:flutter/cupertino.dart';
-import 'package:mawaqit/src/helpers/SharedPref.dart';
-import 'package:mawaqit/src/models/settings.dart';
 import 'package:global_configuration/global_configuration.dart';
 import 'package:http/http.dart';
+import 'package:mawaqit/src/helpers/SharedPref.dart';
+import 'package:mawaqit/src/models/settings.dart';
 
 ValueNotifier<Settings> setting = new ValueNotifier(new Settings());
 
